feat: continue replacing reusable code with composable
